refactor: all providers are inside /layout/providers/index.tsx
