+++ src/app/[locale]/layout.tsx
@@ -3,9 +3,7 @@ import "@/app/globals.css";
 import { ReactNode } from "react";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-
-import Providers from "@/layouts/Providers/Providers";
-import { NextIntlClientProvider, useMessages } from "next-intl";
+import Providers from "@/layouts/Providers";
 
 const inter = Inter({ subsets: ["latin"] });
 export const metadata: Metadata = {
@@ -22,16 +20,12 @@ export default function RootLayout({
   children,
   params: { locale },
 }: RootLayoutProps) {
-  const messages = useMessages();
-
   return (
     <html lang={locale}>
       <body className={inter.className}>
         <Providers>
-          <NextIntlClientProvider messages={messages}>
-            <div className="flex min-h-screen flex-col">{children}</div>
-            <div id="overlay" />
-          </NextIntlClientProvider>
+          <div className="flex min-h-screen flex-col">{children}</div>
+          <div id="overlay" />
         </Providers>
       </body>
     </html>
