--- src/layouts/Providers/Providers.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { ReactNode } from "react";
-import { SessionProvider } from "next-auth/react";
-
-export default function Providers({ children }: { children: ReactNode }) {
-  return <SessionProvider>{children}</SessionProvider>;
-}
