+++ src/layouts/Providers/SessionProvider.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import { ReactNode } from "react";
+import { SessionProvider as NextAuthProvider } from "next-auth/react";
+
+export default function SessionProvider({ children }: { children: ReactNode }) {
+  return <NextAuthProvider>{children}</NextAuthProvider>;
+}
